Decouple eval metric & model and split utils into multiple sources
