+++ examples/hstu/model/ranking_gr.py
@@ -12,7 +12,6 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from collections import OrderedDict
 from typing import Tuple
 
 import torch
@@ -144,24 +143,3 @@ class RankingGR(BaseModel):
             jagged_item_logit.detach(),
             labels.detach(),
         )
-
-    def evaluate_one_batch(self, batch: RankingBatch) -> None:
-        """
-        Evaluate one batch of data.
-
-        Args:
-            batch (RankingBatch): The batch of ranking data.
-        """
-        with torch.no_grad():
-            jagged_item_logit, redistributed_labels = self.get_logit_and_labels(batch)
-            self._metric_module(jagged_item_logit.float(), redistributed_labels)
-
-    def compute_metric(self) -> "OrderedDict":
-        """
-        Compute the evaluation metrics.
-
-        Returns:
-            OrderedDict: The computed metrics.
-        """
-        ret_dict = self._metric_module.compute()
-        return ret_dict
