+++ examples/hstu/pretrain_gr_retrieval.py
@@ -28,11 +28,13 @@ import torch  # pylint: disable-unused-import
 from configs import RetrievalConfig
 from distributed.sharding import make_optimizer_and_shard
 from model import get_retrieval_model
+from modules.metrics import RetrievalTaskMetricWithSampling
 from pipeline.train_pipeline import (
     JaggedMegatronPrefetchTrainPipelineSparseDist,
+    JaggedMegatronTrainNonePipeline,
     JaggedMegatronTrainPipelineSparseDist,
 )
-from utils import (
+from training import (
     NetworkArgs,
     OptimizerArgs,
     TensorModelParallelArgs,
@@ -45,7 +47,6 @@ from utils import (
     get_data_loader,
     get_dataset_and_embedding_args,
     maybe_load_ckpts,
-    train,
     train_with_pipeline,
 )
 
@@ -113,9 +114,11 @@ def main():
         dense_optimizer_param=optimizer_param,
         dynamicemb_options_dict=dynamic_options_dict,
         allowed_compute_kernels=compute_kernels,
-        enable_prefetch_pipeline=trainer_args.enable_prefetch_pipeline,
+        enable_prefetch_pipeline=trainer_args.pipeline_type == "prefetch",
     )
-
+    stateful_metric_module = RetrievalTaskMetricWithSampling(
+        metric_types=task_config.eval_metrics, MAX_K=500
+    )
     train_dataloader, test_dataloader = get_data_loader(
         "retrieval", dataset_args, trainer_args, 0
     )
@@ -131,21 +134,20 @@ def main():
             dense_optimizer,
             device=torch.device("cuda", torch.cuda.current_device()),
         )
-        train_with_pipeline(
-            pipeline,
-            trainer_args,
-            train_dataloader,
-            test_dataloader,
-            dense_optimizer,
-        )
     else:
-        train(
+        pipeline = JaggedMegatronTrainNonePipeline(
             model_train,
-            trainer_args,
-            train_dataloader,
-            test_dataloader,
             dense_optimizer,
+            device=torch.device("cuda", torch.cuda.current_device()),
         )
+    train_with_pipeline(
+        pipeline,
+        stateful_metric_module,  # no stateful_metric_module
+        trainer_args,
+        train_dataloader,
+        test_dataloader,
+        dense_optimizer,
+    )
     init.destroy_global_state()
